Add files via upload
What is your application? 
The application is a system that allows users to and a pool of options regarding meals.
What does it do?
It takes in input from the user (ingredients) and outputs a list of possible meals.
How does it work?
The program employs the use of HashMaps, if  corresponding ingredientes are available then the meal those ingredientes can be made are pushed to the user.  
How will it work on May 24thHow do you want it to work in future after you have more time to develop the idea and your programming skills.
The MVP uses GUI’s to let the user enter ingredients, provide possible meals, as well as provide a shopping list of needed ingredients.